Refactor user entities
