+++ src/entities/Message.java
@@ -1,4 +1,4 @@
-package entity;
+package entities;
 
 import java.time.Instant;
