+++ src/entities/PasswordValidator.java
@@ -1,4 +1,4 @@
-package entity;
+package entities;
 
 public interface PasswordValidator {
   boolean passwordIsValid(String password);
