+++ src/entities/user_entities/AbstractUser.java
@@ -1,0 +1,7 @@
+package entities.user_entities;
+
+public interface AbstractUser {
+  String getUid();
+
+  String getName();
+}
