--- src/entity/User.java
@@ -1,15 +1,0 @@
-package entity;
-
-import java.time.LocalDateTime;
-
-public interface User {
-  String getUid();
-
-  String getEmail();
-
-  String getName();
-
-  String getPassword();
-
-  LocalDateTime getCreationTime();
-}
