--- src/entity/UserFactory.java
@@ -1,8 +1,0 @@
-package entity;
-
-import java.time.LocalDateTime;
-
-public interface UserFactory {
-  /** Requires: password is valid. */
-  User create(String uid, String email, String name, String password, LocalDateTime ltd);
-}
