+++ src/use_case/chat/ChatInteractor.java
@@ -1,6 +1,6 @@
 package use_case.chat;
 
-import entity.Message;
+import entities.Message;
 import java.time.Instant;
 
 public class ChatInteractor implements ChatInputBoundary {
