+++ src/use_case/chat/ChatOutputBoundary.java
@@ -1,6 +1,6 @@
 package use_case.chat;
 
-import entity.Message;
+import entities.Message;
 import java.util.List;
 
 public interface ChatOutputBoundary {
