+++ src/use_case/search/SearchOutputData.java
@@ -1,6 +1,6 @@
 package use_case.search;
 
-import entity.SearchReponseArray;
+import entities.SearchReponseArray;
 
 public class SearchOutputData {
   private final SearchReponseArray response;
