+++ src/use_case/signup/SignupInteractor.java
@@ -1,7 +1,7 @@
 package use_case.signup;
 
-import entity.User;
-import entity.UserFactory;
+import entities.user_entities.User;
+import entities.user_entities.UserFactory;
 import java.time.LocalDateTime;
 import java.util.Optional;
