+++ src/use_case/signup/SignupUserDataAccessInterface.java
@@ -1,6 +1,6 @@
 package use_case.signup;
 
-import entity.User;
+import entities.user_entities.User;
 import java.util.Optional;
 
 public interface SignupUserDataAccessInterface {
